pg_trgm: fix train-6 audit blockers B1/B2 + suite riders
B1: compact_trigram now uses crc32c::legacy_crc32_lexeme (the reflected-table
LEGACY_CRC32 of pg_crc.h, trgm_op.c:341) instead of the zlib-variant
traditional_crc32 — hashed (multibyte) trigram bytes now match C, making
show_trgm/similarity C-identical and index keys compatible.
B2: operator threshold comparisons promote the float4 similarity to double
(C's comparison type) instead of demoting the GUC to f32 — exact-boundary
values now match C (incl. the word-similarity CHECK_ONLY early exit).
Riders in sql/pg_utf8_trgm.sql make both classes suite-visible via
contrib-e2e's live-C diff.

+++ crates/contrib/pg_trgm/src/trgm.rs
@@ -448,7 +448,7 @@ fn iterate_word_similarity(
                         lower = tmp_lower;
                         count = tmp_count;
                     }
-                    if check_only && smlr_cur >= threshold as f32 {
+                    if check_only && smlr_cur as f64 >= threshold {
                         break;
                     }
                 }
@@ -464,7 +464,7 @@ fn iterate_word_similarity(
             }
 
             smlr_max = smlr_max.max(smlr_cur);
-            if check_only && smlr_max >= threshold as f32 {
+            if check_only && smlr_max as f64 >= threshold {
                 break;
             }
